51cto-automator/requirements.txt
adbutils==0.11.0
alabaster==0.7.12
allure-pytest==2.9.45
allure-python-commons==2.9.45
apkutils2==1.0.0
Appium-Python-Client==1.2.0
attrs==21.2.0
Babel==2.9.1
bcrypt==3.2.0
cached-property==1.5.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
cigam==0.0.3
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
dataclasses==0.8
decorator==5.0.9
Deprecated==1.2.12
deprecation==2.1.0
docopt==0.6.2
docutils==0.17.1
filelock==3.4.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Foundations==2.0.7
greenlet==1.1.2
idna==2.10
imagesize==1.3.0
importlib-metadata==4.0.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
linecache2==1.0.0
logzero==1.5.0
lxml==4.6.3
Manager==2.0.3
MarkupSafe==2.0.1
numpy==1.19.5
ordereddict==1.1
packaging==20.9
paramiko==2.9.1
Pillow==8.1.0
pipreqs==0.4.11
pluggy==0.13.1
progress==1.5
py==1.10.0
pyaml==20.4.0
pycparser==2.21
pyelftools==0.27
Pygments==2.11.1
PyNaCl==1.4.0
pyparsing==2.4.7
PySnooper==1.0.0
pytest==6.2.4
pytest-assume==2.4.3
pytest-html==3.1.1
pytest-metadata==1.11.0
pytz==2021.3
PyYAML==5.4.1
rarfile==4.0
requests==2.25.1
retry==0.9.2
selenium==3.141.0
six==1.16.0
snowballstemmer==2.2.0
some-package==0.1
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.29
toml==0.10.2
traceback2==1.4.0
typing-extensions==3.10.0.0
uiautomator2==2.16.13
unittest2==1.1.0
urllib3==1.26.4
Werkzeug==2.0.2
whichcraft==0.6.1
wrapt==1.12.1
xlrd==1.2.0
xmltodict==0.12.0
yarg==0.1.9
zipp==3.4.1
